Initial Commit
The main executable is created mandelbrot.c - it must be linked with libpng, using
the -lpng option via gcc. The building is all automated using make (see the associated
Makefile).

In addition, an Image ADT was created. Essentially, this is a wrapper around
libpng that allows you to create an image of static dimensions and then set
the pixel values in RGB. The created image can be saved to the disk.