Optimize code slightly with loop unrolling and function inlining
The iterations loop was unrolled to half its size - this saves
(width * height * iterations) / 2 iterations, which is pretty significant considering
that the entire loop takes (width * height * iterations) otherwise.

In addition, the crpow function is now inlined and a faster while loop is used to
iteratively calculate the complex exponentiation.

Finally, the pixel-to-Cartesian coordinate mapping was simplified by using a
repeated addition instead of multiplication.