Separate generate_mandelbrot_set from mandelbrot.c to allow for simpler builds
Instead of building a single binary executable, two are built with the C and x86
versions of generate_mandelbrot_set, respectively. As such, the x86 version
was renamed to generate_mandelbrot_set-x86.s.

The generate_mandelbrot_set interface is described in mandelbrot.h, which
is included by mandelbrot.c and implemented by generate_mandelbrot_set.c.