Finished first draft of translating generate_mandelbrot_set to x86
This draft is mostly complete, but barely tested. Namely, there is a segfault
on the call of Image_setPixel.

In addition, the iteration order was reversed (for each for loop) in
generate_mandelbrot_set to be more intuitive when translating to x86 - it is
easier to check if something is zero than anything else.