Fix x86 generate_mandelbrot_set issue of comparisons
As expected, the issue lay in using comisd - this was not the appropriate comparison
mneumonic. Instead, we wanted to use cmplesd, which compares two scalar doubles
for the destination being less than or equal to the source.

The result of the comparison, however, is stored in the destination, and so we have
a few more operations to actually make use of the comparison data.

The x86 is yet to be optimized fully because it still takes a significant amount of time -
almost 3x as much as the C version, even when using the -O3 compilation flag.
Most likely, the overhead is introduced from the massive number of data movement
needed before/after each call of _Image_setPixel.