Fixed primary timing issue in x86
The main issue came with determing the number of iterations - we are using a
left shift, which doubles the number of iterations. Instead, we should use a right
shift, which halves it.

In addition, a few operations were changed to using the single double variant,
which should be faster (in theory) then adding both doubles.