Roll iteration loop in C version of generate_mandelbrot_set to prevent overflow
When the loop was compressed into two virtual iterations per physical iteration,
there were two places for error. However, if the first call of crpow overflowed,
the second might make it positive again. Thus, this let certain errors go undetected.
In particular, it let a region directly around the Mandelbrot Set for high exponents
(about 35) to be marked as in the set, even though it was not. Re-rolling the loop
prevented this issue.

It is worth noting that the overflow is detected internally in the x86 version, and thus
the bug does not occur even with the loop unrolled.