Use both high/low portions of the xmm registers in x86
This allows for more compact operations, especially for associative mathematical
operations. For example, if we need to multiply the two real parts and the two
imaginary parts, respectively, this can be done with a single operation if the real
parts are both in the high quad and the imaginary are in the low word (or vice versa).

Currently, the program runs into an infinite loop somewhere and this has to be fixed.