Fix x86 packed double issue of drawing a rectangle
The primary issue came with the use of saving the current registers.
movq only copies the lower 64 bits of the xmm register (i.e the low quad word)
into the quad word register. Thus, the high quad word of the xmm registers were
not being saved - movq might even set the upper quad word to 0, thus causing
a weird mess of issues.

To prevent this, we switch to using callee-saved registers to store the nine primary
doubles. As a result, some of the scratch registers used were also changed
accordingly.