Optimize by using packed doubles to compute in parallel
Essentially, two complex numbers are stored in the register (high and low doubles).
The second complex number is just past the first one. Thus, we move two steps
in each width iteration, and decrement the width by two.

To keep track of which points to draw, the bl and bh registers are used. When both
are false, the iterations stop. If the iterations complete before both are false,
then the points that should be drawn are drawn to the image - this can be one or
both of the points.

This speeds up the computations by about 40%. It is not a perfect halved time
because a point may be out of bounds before the other, but the iterations must
continue until either both are out of bounds or the iterations are completed.
Otherwise, managing a separate iteration count for each would be much more
challenging.

To ensure that the computations work properly (and do not run off the bounds of
the pixels array), the width and height are enforced to be even integers in
mandelbrot.c.

Note: there is a small bug currently with drawing the points. In some cases with large
iterations, a pixel or two are not drawn when they should be - I'm not exactly sure
why this occurs.