Create imgdiff tool to take the difference of two images as a count
Using this, we can obtain a more accurate difference reporting when comparing
the two resulting images from the C and x86 versions.

To do so, a few methods were added to the Image_T ADT, the primary being the
Image_fromFile method, which reads an image from a PNG file.